% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\docType{data}
\name{.type_colours}
\alias{.type_colours}
\title{Palette constants}
\format{
An object of class \code{character} of length 8.
}
\usage{
.type_colours
}
\description{
Shared colour definitions used by all plot functions.
}
\keyword{internal}
